splitting more components into functions for easier management
